Expect python-carbon to provide carbon-cache systemd service
The calamari-carbon.service file really didn't belong to calamari - it
starts part of carbon, so it belongs in the python-carbon package.  On
SUSE, at least, this service expects carbon config to be in
/etc/carbon/carbon.conf, not /etc/graphite/carbon.conf, hence the change
to the SUSE spec file.  The Makefile still installs carbon.conf in
/etc/graphite/ as I'm not sure what other distros want to do.

This commit also relinquishes calamari-server's ownership of some
/var/lib/graphite directories that are owned by the graphite-web
package.

Signed-off-by: Tim Serong <tserong@suse.com>