Move salt files from /opt/calamari/salt to /srv/
This moves /opt/calamari/salt/{pillar,reactor,salt} to /srv/, which is
the directory hierarchy used in the examples listed in /etc/salt/master
on SUSE.

/opt/calamari/salt-local/ has been moved to /usr/share/calamari/salt/,
as these files are only used internally by Calamari to configure the
server.

Note that this change will affect all distros, so shouldn't be pushed
upstream at this time.

Signed-off-by: Tim Serong <tserong@suse.com>