cthulhu: fix for salt 2016: lower poll timeout, add gevent.idle()
Somehow, when using salt 2016, it seems the ClusterMonitor,
TopLevelEvents and SeverMonitor greenlets starve everything else,
resulting in "Lost remote after 10s heartbeat" in the logs, and
"Server Error (500)" presented to users.

This commit makes each of the above call gevent.idle() in their
respective loops, and reduces the POLL_TIMEOUT in salt_wrapper
from 5 seconds to 1 second, which seems to grease the wheels enough
that everything works again.

I'm not entirely happy with this, as I've not been able to determine
the root cause of the problem, but, it works :-/

Signed-off-by: Tim Serong <tserong@suse.com>